e2e: use gpg to run key creation
