+++ .test.pub
@@ -1,0 +1,1 @@
+ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAID+i0E1KeXDlaHRYEqh7TRidqhJjEBF+qVdS75YaCQoo emma.doyle@EmmaDoyleMBP
